Issue #11 Set ClientIp via the parameters form.
Still not ideal - seems too much repitition in code. Also the clientIp is
not retreived from the session, but many other fields (e.g. return URLs) are also
not retrieved from the session when re-presenting the request form, so it is
consistent at least, but could do with some improvements.